vue2版本的capture
