+++ packages/app-backend/src/engine/vue2/vue2.ts
@@ -1,6 +1,6 @@
 // API vue2
 import { camelize, getComponentName, getCustomRefDetails } from '@utils/util'
-import { functional } from './capture'
+import { functional, capture } from './capture'
 
 function getInstanceName(instance) {
   const name = getComponentName(instance.$options || instance.fnOptions || {})
@@ -29,4 +29,5 @@ export default {
   findComponentByEl,
   isFragment,
   functional,
+  capture,
 }
