+++ packages/app-backend/src/engine/vue3/vue3.ts
@@ -27,4 +27,5 @@ export default {
   getInstanceName,
   findComponentByEl,
   isFragment,
+  capture: () => {},
 }
